allow override blocks for default actions

+++ lib/active_admin_scoped_collection_actions/dsl.rb
@@ -14,9 +14,8 @@ module ActiveAdminScopedCollectionActions
       config.add_scoped_collection_action(name, options)
     end
 
-
-    def add_scoped_collection_action_default_update(name, options, &block)
-      batch_action :batch_update, if: proc { false } do |selection, _|
+    def add_scoped_collection_action_default_update(options, &block)
+      batch_action :scoped_collection_update, if: proc { false } do |selection, _|
         collection = selection.any? ? resource_class.where(id: selection) : batch_action_collection
         unless authorized?(:batch_edit, resource_class)
           flash[:error] = 'Access denied'
@@ -25,10 +24,14 @@ module ActiveAdminScopedCollectionActions
         if !params.has_key?(:changes) || params[:changes].empty?
           render nothing: true, status: :no_content and next
         end
-        permitted_changes = params.require(:changes).permit( *(options[:form].call.keys) )
+        permitted_changes = params.require(:changes).permit(*(options[:form].call.keys))
         errors = []
-        collection.find_each do |record|
-          errors << "#{record.id} | #{record.errors.full_messages.join('. ')}" unless update_resource(record, [permitted_changes])
+        if block_given?
+          instance_eval &block
+        else
+          collection.find_each do |record|
+            errors << "#{record.id} | #{record.errors.full_messages.join('. ')}" unless update_resource(record, [permitted_changes])
+          end
         end
         if errors.empty?
           flash[:notice] = 'Batch update done'
@@ -39,17 +42,20 @@ module ActiveAdminScopedCollectionActions
       end
     end
 
-
-    def add_scoped_collection_action_default_destroy(name, _, &block)
-      batch_action :batch_destroy, if: proc { false } do |selection|
+    def add_scoped_collection_action_default_destroy(_, &block)
+      batch_action :scoped_collection_destroy, if: proc { false } do |selection|
         collection = selection.any? ? resource_class.where(id: selection) : batch_action_collection
         unless authorized?(:batch_destroy, resource_class)
           flash[:error] = 'Access denied'
           render nothing: true, status: :no_content and next
         end
         errors = []
-        collection.find_each do |record|
-          errors << "#{record.id} | Cant be destroyed}" unless destroy_resource(record)
+        if block_given?
+          instance_eval &block
+        else
+          collection.find_each do |record|
+            errors << "#{record.id} | Cant be destroyed}" unless destroy_resource(record)
+          end
         end
         if errors.empty?
           flash[:notice] = 'Batch destroy done'
